Moved the wait_for_sshd inside the keypair/password/fixed password logic
The wait_for_sshd should honour the chosen authentication method, and only
try that specific method.

PS: This fix isn't perfect yet, as it still prompts for a password when the
provided password isn't accepted (can happen if SSH is already up, but
the servers hasnt updated the users password through the cloud-utils yet).
I have a solution for that too, but it depends on changes in the test-kitchen.